fix: brand home link missing styles on non-home pages
It wasn't clear according to the docs
(@see https://tanstack.com/router/v1/docs/guide/navigation#link-component)
that `activeProps` were for props that will be added to the component
only when the `<Link>` path matches the current URL path. A little silly
on my part for missing that, because I feel like it's pretty self-explanatory.
However, the documentation assumes that you know to use the `className` prop
and didn't include it.. so that's where I got confused and incorrectly
assumed that you can't use `className`, so I thought passing the `className`
through the `activeProps` props was the only way.

Signed-off-by: Jack Ta <jack.th.ta@gmail.com>